+++ lib/StringUtils.lua
@@ -1,6 +1,6 @@
 local StringUtils = {}
 
-function StringUtils.Split(str, delimiter) --: ({string, number}) => string
+function StringUtils.Split(str, delimiter) --: (string, number) => string
 	local result = {}
 	local from = 1
 	local delim_from, delim_to = string.find(str, delimiter, from)
@@ -15,6 +15,14 @@ end
 
 function StringUtils.Trim(s) --: (string) => string
 	return s:match "^%s*(.-)%s*$"
+end
+
+function StringUtils.StartsWith(str, start) --: (string, string) => boolean
+	return string.sub(str, 1, string.len(start)) == start
+end
+
+function StringUtils.EndsWith(str, ending) --: (string, string) => boolean
+	return string.sub(str, -string.len(ending)) == ending
 end
 
 return StringUtils