+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * `StringUtils.Split`
 * `StringUtils.Trim`
+* `StringUtils.StartsWith`
+* `StringUtils.EndsWith`
 * `TableUtils.Slice`
 * `TableUtils.GetLength`
 * `TableUtils.Clone`